fix: have hydration interspersed with crawl output
Previously, the crawl output was using one rich.Progress bar for
all tasks - which meant the Live region was active when the hydration
tasks were printing their reports. And thus the report got punted to
above the progress bars. This was extremely confusing output.

So instead of trying to be clever and reuse the same progress bar
for all crawling tasks, just make a new one each time. This lets
the output be appropriately interwoven.

Fixes: https://github.com/smart-on-fhir/smart-fetch/issues/20