crawl: stop using clinical dates as fallbacks for admin dates
When using --since-mode=created, we had been using clinical dates
for resources that didn't have admin dates.

But in practice, that seems fraught. Both because external sources
can inject some resources and lightly-backdated resources might be
missed around the cutoff date.

Since the affected resources are not huge by volume usually, we'll
just grab all of them each time (like we do for Device and Patient
already).

Also, this commit will ignore any potential transaction time dates
that are in the future, when examining the data (these are likely
mistakes) - but I saw this happen in real data.