Be more strict about which symlinks we create
Backstory:
For managed directories, we maintain symlinks into the subfolders for
the convience of anyone working with "the export as a whole" - each
subfolder still holds its log files and the real data, but there's also
a "world wide view" offered by the symlinks. So users can pick and
choose which approach makes sense for them.

(Like, in Cumulus land, you'll want to ETL the subdirs directly, but
maybe if you're examining data, you want to just grep through the
symlinks.)

Previous situation:
Previously, we would never remove a symlink and added to the collection
after every export. This makes sense when incrementally building a pile
of ndjson, but makes less sense once you start doing new full exports.
All of a sudden, your symlinks point at twice as much data as
necessary, including potentially-deleted rows.

New situation:
We now try harder to notice that old obsolete data does not need to be
pointed at. We follow chains of resources back to the most recent full
export. Though there are still some edge cases where we might be more
generous than we need to, because we aren't examining the actual
content of ndjson files:
- we don't notice duplicated hydrated resources from multiple full
  exports - so you'll only save space there if your full exports
  include all possible sources of the hydrated resource
- if you are using type filters heavily, we treat each "chain" of
  type filters as distinct - so if you exported conditions with a
  certain code and later exported conditions with a certain severity,
  we won't notice if there are duplicates. But we would ignore older
  full severity exports and only include the new ones.

Also, as an aid for folks that are maybe moving export archives around,
to or from places that don't support symlinks (like Amazon S3), there's
now a new "reset-symlinks" subcommand to recreate all symlinks as
needed. I'd like to add a more comprehensive solution later, like a
"sync" subcommand maybe, but this is a nice escape valve in the
meantime and for folks who accidentally messed up their symlinks.