Add new 'bundle' subcommand and --bundle args for crawl/bulk
All these do a similar thing: turn a folder of NDJSON data into a
single Bundle.json.gz file that holds all the data.

This is useful for some tools that like to see a single Bundle.

So whether you have an existing pile of data you want to convert or
are getting your data direct from smart-fetch, you can now have it
as a Bundle.

Because the Bundle format is not as convenient to work with, we
don't currently support reading it as input for subcommands like
'hydrate' or allow its use in 'export' managed directories. It's
more of an end-state format, for other tools to consume.