hydrate: speed up PractitionerRole searching on second time
If interrupted and we need to re-hydrate PractitionerRoles, first look
to see which Practitioners we already searched for.

This works by separating out the searched Roles from the manually
referenced Roles, so we know which is which.

Also, add a little more output to hydration tasks, to make it clearer
what's happening.